Moved rule to CleanCode section.
