+++ src/main/resources/rulesets/design.xml
@@ -225,28 +225,4 @@ class SuspectCode {
             ]]>
         </example>
     </rule>
-
-    <rule name="DuplicatedArrayKey"
-          message="Duplicated array entry for key {0}, first declared at line {1}."
-          class="PHPMD\Rule\Design\DuplicatedArrayKey"
-          externalInfoUrl="http://phpmd.org/rules/design.html#duplicatedarraykey">
-        <description>
-            <![CDATA[
-Defining another value for the same key in array literal overrides previous one,
-which makes it effectively an unused code. If it's known from the beginning that
-the key will have different value, there is usually no point in defining first one.
-            ]]>
-        </description>
-        <priority>2</priority>
-        <example>
-            <![CDATA[
-function createArray() {
-    return [
-        'foo' => 'bar', // this one makes no effect
-        "foo" => 'baz',
-    ];
-}
-            ]]>
-        </example>
-    </rule>
 </ruleset>
