+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesMultipleTimesToFunctionWithAssotiativeArrayDefinitionWithDuplicatedKeys.php
