+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesMultipleTimesToMethodWithAssotiativeArrayDefinitionWithDuplicatedKeys.php
