+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesToFunctionWithAssotiativeArrayDefinitionWithDuplicatedKeys.php
