+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesToFunctionWithAssotiativeArrayDefinitionWithDuplicatedMixedQuotedKeys.php
