+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesToFunctionWithAssotiativeArrayDefinitionWithDuplicatedMixedTypeKeys.php
