+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesToMethodWithAssotiativeArrayDefinitionWithDuplicatedKeys.php
