+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesToMethodWithAssotiativeArrayDefinitionWithDuplicatedMixedQuotedKeys.php
