+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleAppliesToMethodWithAssotiativeArrayDefinitionWithDuplicatedMixedTypeKeys.php
