+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleNotAppliesToFunctionWithAssotiativeArrayDefinitionWithoutDuplicatedKeys.php
