+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleNotAppliesToFunctionWithNonAssotiativeArrayDefinition.php
