+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleNotAppliesToFunctionWithoutArrayDefinition.php
