+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleNotAppliesToMethodWithAssotiativeArrayDefinitionWithoutDuplicatedKeys.php
