+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleNotAppliesToMethodWithNonAssotiativeArrayDefinition.php
