+++ src/test/resources/files/Rule/CleanCode/DuplicatedArrayKey/testRuleNotAppliesToMethodWithoutArrayDefinition.php
